+++ lib/rubocop/config.rb
@@ -55,6 +55,12 @@ module RuboCop
       'Layout/FirstParameterIndentation' =>
         'The `Layout/FirstParameterIndentation` cop has been renamed to ' \
         '`Layout/IndentFirstArgument`.',
+      'Layout/IndentArray' =>
+        'The `Layout/IndentArray` cop has been renamed to ' \
+        '`Layout/IndentFirstArrayElement`.',
+      'Layout/IndentHash' =>
+        'The `Layout/IndentHash` cop has been renamed to ' \
+        '`Layout/IndentFirstHashElement`.',
       'Layout/SpaceAfterControlKeyword' =>
         'The `Layout/SpaceAfterControlKeyword` cop has been removed. Please ' \
         'use `Layout/SpaceAroundKeyword` instead.',